@@ -1,34 +1,30 @@
 module github.com/OpenSlides/openslides-projector-service
 
-go 1.23.6
+go 1.24.0
 
 require (
-	github.com/OpenSlides/openslides-go v0.0.0-20250208121842-319b8084179d
+	github.com/OpenSlides/openslides-go v0.0.0-20250521145008-8e4ed1371a71
 	github.com/caarlos0/env/v6 v6.10.1
-	github.com/goccy/go-yaml v1.15.23
-	github.com/iancoleman/strcase v0.3.0
-	github.com/jackc/pgx/v5 v5.7.2
-	github.com/redis/go-redis/v9 v9.7.1
+	github.com/goccy/go-yaml v1.17.1
 	github.com/rs/zerolog v1.33.0
-	golang.org/x/text v0.23.0
 )
 
 require (
-	github.com/golang-jwt/jwt/v4 v4.5.1 // indirect
+	github.com/golang-jwt/jwt/v4 v4.5.2 // indirect
 	github.com/gomodule/redigo v1.9.2 // indirect
 	github.com/ostcar/topic v0.4.1 // indirect
 )
 
 require (
-	github.com/cespare/xxhash/v2 v2.3.0 // indirect
-	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
 	github.com/jackc/pgpassfile v1.0.0 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20240606120523-5a60cdf6a761 // indirect
+	github.com/jackc/pgx/v5 v5.7.4 // indirect
 	github.com/jackc/puddle/v2 v2.2.2 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/stretchr/testify v1.9.0 // indirect
-	golang.org/x/crypto v0.31.0 // indirect
-	golang.org/x/sync v0.12.0 // indirect
-	golang.org/x/sys v0.28.0 // indirect
+	golang.org/x/crypto v0.38.0 // indirect
+	golang.org/x/sync v0.14.0 // indirect
+	golang.org/x/sys v0.33.0 // indirect
+	golang.org/x/text v0.25.0 // indirect
 )
