@@ -4,22 +4,54 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"html/template"
 	"slices"
 	"strconv"
+	"time"
 
+	"github.com/OpenSlides/openslides-go/datastore/dsfetch"
+	"github.com/OpenSlides/openslides-go/datastore/dsmodels"
+	"github.com/OpenSlides/openslides-go/datastore/flow"
 	"github.com/OpenSlides/openslides-projector-service/pkg/database"
-	"github.com/OpenSlides/openslides-projector-service/pkg/models"
 	"github.com/OpenSlides/openslides-projector-service/pkg/projector/slide"
 	"github.com/rs/zerolog/log"
 )
+
+type ProjectorSettings struct {
+	MeetingName            string
+	MeetingDescription     string
+	MeetingLogo            int
+	HeaderImage            int
+	Name                   string
+	IsInternal             bool
+	Scale                  int
+	Scroll                 int
+	Width                  int
+	AspectRatioNumerator   int
+	AspectRatioDenominator int
+	Color                  string
+	BackgroundColor        string
+	HeaderBackgroundColor  string
+	HeaderFontColor        string
+	HeaderH1Color          string
+	ChyronBackgroundColor  string
+	ChyronBackgroundColor2 string
+	ChyronFontColor        string
+	ChyronFontColor2       string
+	ShowHeaderFooter       bool
+	ShowTitle              bool
+	ShowLogo               bool
+	ShowClock              bool
+}
 
 type projector struct {
 	ctxCancel      context.CancelFunc
 	db             *database.Datastore
 	slideRouter    *slide.SlideRouter
-	projector      *models.Projector
+	projector      *dsmodels.Projector
+	pSettings      *ProjectorSettings
 	listeners      []chan *ProjectorUpdateEvent
 	Content        string
 	Projections    map[int]template.HTML
@@ -32,30 +64,24 @@ type ProjectorUpdateEvent struct {
 	Data  string
 }
 
-func newProjector(parentCtx context.Context, id int, db *database.Datastore) (*projector, error) {
-	projectorQuery := database.Collection(db, &models.Projector{}).SetIds(id)
-	data, err := projectorQuery.GetOne()
+func newProjector(parentCtx context.Context, id int, db *database.Datastore, ds flow.Flow) (*projector, error) {
+	ctx, cancel := context.WithCancel(parentCtx)
+
+	data, err := db.Fetch.Projector(id).First(ctx)
 	if err != nil {
+		cancel()
 		return nil, fmt.Errorf("error fetching projector from db %w", err)
 	}
 
-	if data == nil {
-		return nil, fmt.Errorf("projector not found")
-	}
-
-	ctx, cancel := context.WithCancel(parentCtx)
 	p := &projector{
 		ctxCancel:      cancel,
 		db:             db,
-		projector:      data,
-		slideRouter:    slide.New(ctx, db),
+		projector:      &data,
+		pSettings:      &ProjectorSettings{},
+		slideRouter:    slide.New(ctx, db, ds),
 		Projections:    make(map[int]template.HTML),
 		AddListener:    make(chan chan *ProjectorUpdateEvent),
 		RemoveListener: make(chan (<-chan *ProjectorUpdateEvent)),
-	}
-
-	if err = p.updateFullContent(); err != nil {
-		return nil, fmt.Errorf("error generating projector content: %w", err)
 	}
 
 	go p.subscribeProjector(ctx)
@@ -80,14 +106,64 @@ func newProjector(parentCtx context.Context, id int, db *database.Datastore) (*p
 
 func (p *projector) subscribeProjector(ctx context.Context) {
 	defer p.ctxCancel()
-	// TODO: Subscribe on projector settings updates
-	// Ignore e.g. projection defaults and [...]_projection_ids
-	// If header active: Meeting name + description need to be subscribed
-	projectorSub, err := database.Collection(p.db, &models.Projector{}).SetIds(p.projector.ID).SubscribeOne(p.projector)
-	if err != nil {
-		log.Fatal().Err(err).Msg("could not open projector subscription")
-	}
-	defer projectorSub.Unsubscribe()
+	p.db.NewContext(ctx, func(f *dsmodels.Fetch) {
+		f.Projector_Name(p.projector.ID).Lazy(&p.pSettings.Name)
+		f.Projector_IsInternal(p.projector.ID).Lazy(&p.pSettings.IsInternal)
+		f.Projector_Scale(p.projector.ID).Lazy(&p.pSettings.Scale)
+		f.Projector_Scroll(p.projector.ID).Lazy(&p.pSettings.Scroll)
+		f.Projector_Width(p.projector.ID).Lazy(&p.pSettings.Width)
+		f.Projector_AspectRatioNumerator(p.projector.ID).Lazy(&p.pSettings.AspectRatioNumerator)
+		f.Projector_AspectRatioDenominator(p.projector.ID).Lazy(&p.pSettings.AspectRatioDenominator)
+		f.Projector_Color(p.projector.ID).Lazy(&p.pSettings.Color)
+		f.Projector_BackgroundColor(p.projector.ID).Lazy(&p.pSettings.BackgroundColor)
+		f.Projector_HeaderBackgroundColor(p.projector.ID).Lazy(&p.pSettings.HeaderBackgroundColor)
+		f.Projector_HeaderFontColor(p.projector.ID).Lazy(&p.pSettings.HeaderFontColor)
+		f.Projector_HeaderH1Color(p.projector.ID).Lazy(&p.pSettings.HeaderH1Color)
+		f.Projector_ChyronBackgroundColor(p.projector.ID).Lazy(&p.pSettings.ChyronBackgroundColor)
+		f.Projector_ChyronBackgroundColor2(p.projector.ID).Lazy(&p.pSettings.ChyronBackgroundColor2)
+		f.Projector_ChyronFontColor(p.projector.ID).Lazy(&p.pSettings.ChyronFontColor)
+		f.Projector_ChyronFontColor2(p.projector.ID).Lazy(&p.pSettings.ChyronFontColor2)
+		f.Projector_ShowHeaderFooter(p.projector.ID).Lazy(&p.pSettings.ShowHeaderFooter)
+		f.Projector_ShowTitle(p.projector.ID).Lazy(&p.pSettings.ShowTitle)
+		f.Projector_ShowLogo(p.projector.ID).Lazy(&p.pSettings.ShowLogo)
+		f.Projector_ShowClock(p.projector.ID).Lazy(&p.pSettings.ShowClock)
+		f.Meeting_Name(p.projector.MeetingID).Lazy(&p.pSettings.MeetingName)
+		f.Meeting_Description(p.projector.MeetingID).Lazy(&p.pSettings.MeetingDescription)
+		var logo dsfetch.Maybe[int]
+		f.Meeting_LogoProjectorMainID(p.projector.ID).Lazy(&logo)
+		var header dsfetch.Maybe[int]
+		f.Meeting_LogoProjectorHeaderID(p.projector.ID).Lazy(&header)
+
+		err := f.Execute(ctx)
+		var doesNotExist dsfetch.DoesNotExistError
+		if errors.As(err, &doesNotExist) {
+			p.sendToAll(&ProjectorUpdateEvent{"deleted", ""})
+			p.ctxCancel()
+			return
+		} else if err != nil {
+			log.Error().Err(err).Msg("failed to update projector data")
+			return
+		}
+
+		if val, set := logo.Value(); set {
+			p.pSettings.MeetingLogo = val
+		}
+
+		if val, set := header.Value(); set {
+			p.pSettings.HeaderImage = val
+		}
+
+		encodedData, err := json.Marshal(p.pSettings)
+		if err != nil {
+			log.Error().Err(err).Msg("could not encode projector data")
+		} else {
+			p.sendToAll(&ProjectorUpdateEvent{"settings", string(encodedData)})
+		}
+
+		if err = p.updateFullContent(); err != nil {
+			log.Error().Err(err).Msg("error generating projector content after settings update")
+		}
+	})
 
 	projectionUpdate, projections, err := p.getProjectionSubscription(ctx)
 	if err != nil {
@@ -102,7 +178,7 @@ func (p *projector) subscribeProjector(ctx context.Context) {
 			p.listeners = append(p.listeners, listener)
 			listener <- &ProjectorUpdateEvent{
 				Event: "connected",
-				Data:  "",
+				Data:  strconv.Itoa(int(time.Now().Unix())),
 			}
 		case listener := <-p.RemoveListener:
 			i := slices.IndexFunc(p.listeners, func(el chan *ProjectorUpdateEvent) bool { return el == listener })
@@ -110,26 +186,6 @@ func (p *projector) subscribeProjector(ctx context.Context) {
 				close(p.listeners[i])
 				p.listeners[i] = p.listeners[len(p.listeners)-1]
 				p.listeners = p.listeners[:len(p.listeners)-1]
-			}
-		case updatedFields, ok := <-projectorSub.Channel:
-			if !ok {
-				p.sendToAll(&ProjectorUpdateEvent{"deleted", ""})
-				return
-			}
-
-			updatedData := map[string]interface{}{}
-			for _, field := range updatedFields {
-				updatedData[field] = p.projector.Get(field)
-			}
-			encodedData, err := json.Marshal(updatedData)
-			if err != nil {
-				log.Error().Err(err).Msg("could not encode projector data")
-			} else {
-				p.sendToAll(&ProjectorUpdateEvent{"settings", string(encodedData)})
-			}
-
-			if err = p.updateFullContent(); err != nil {
-				log.Error().Err(err).Msg("error generating projector content after settings update")
 			}
 		case data, ok := <-projectionUpdate:
 			if !ok {
@@ -189,8 +245,8 @@ func (p *projector) updateFullContent() error {
 	}
 
 	var content bytes.Buffer
-	err = tmpl.Execute(&content, map[string]interface{}{
-		"Projector":   p.projector,
+	err = tmpl.Execute(&content, map[string]any{
+		"Projector":   p.pSettings,
 		"Projections": p.Projections,
 	})
 	if err != nil {
@@ -207,42 +263,43 @@ func (p *projector) getProjectionSubscription(ctx context.Context) (<-chan []int
 	projections := make(map[int]string)
 	addProjection := make(chan int)
 	removeProjection := make(chan int)
-	var projectionIDs []int
-	sub, err := database.Collection(p.db, &models.Projector{}).SetIds(p.projector.ID).SetFields("current_projection_ids").SubscribeField(&projectionIDs)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to subscibe projection ids: %w", err)
-	}
 
+	projectionChannel := p.slideRouter.SubscribeContent(addProjection, removeProjection)
 	go func() {
-		defer sub.Unsubscribe()
 		defer close(updateChannel)
 		defer close(addProjection)
 		defer close(removeProjection)
 
-		projectionChannel := p.slideRouter.SubscribeContent(addProjection, removeProjection)
+		p.db.NewContext(ctx, func(f *dsmodels.Fetch) {
+			projectionIDs, err := f.Projector_CurrentProjectionIDs(p.projector.ID).Value(ctx)
+			if err != nil {
+				log.Error().Err(err).Msg("failed to subscibe projection ids")
+			}
+
+			updated := []int{}
+			for id := range projections {
+				if !slices.Contains(projectionIDs, id) {
+					updated = append(updated, id)
+					removeProjection <- id
+					delete(projections, id)
+				}
+			}
+
+			for _, id := range projectionIDs {
+				if _, ok := projections[id]; !ok {
+					addProjection <- id
+				}
+			}
+
+			if len(updated) > 0 || len(projectionIDs) == 0 {
+				updateChannel <- updated
+			}
+		})
+
 		for {
 			select {
 			case <-ctx.Done():
 				return
-			case <-sub.Channel:
-				updated := []int{}
-				for id := range projections {
-					if !slices.Contains(projectionIDs, id) {
-						updated = append(updated, id)
-						removeProjection <- id
-						delete(projections, id)
-					}
-				}
-
-				for _, id := range projectionIDs {
-					if _, ok := projections[id]; !ok {
-						addProjection <- id
-					}
-				}
-
-				if len(updated) > 0 || len(projectionIDs) == 0 {
-					updateChannel <- updated
-				}
 			case update := <-projectionChannel:
 				projections[update.ID] = update.Content
 				updateChannel <- []int{update.ID}
