@@ -1,17 +1,11 @@
 package slide
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"html/template"
-	"reflect"
-	"strings"
 
-	"github.com/OpenSlides/openslides-projector-service/pkg/database"
-	"github.com/OpenSlides/openslides-projector-service/pkg/models"
-	"github.com/rs/zerolog/log"
+	"github.com/OpenSlides/openslides-projector-service/pkg/viewmodels"
 )
 
 type currentSpeakerChyronSlideOptions struct {
@@ -19,166 +13,90 @@ type currentSpeakerChyronSlideOptions struct {
 	AgendaItem bool   `json:"agenda_item"`
 }
 
-func CurrentSpeakerChyronSlideHandler(ctx context.Context, req *projectionRequest) (<-chan string, error) {
-	content := make(chan string)
-	projection := req.Projection
-
-	referenceProjectorId := 0
-	refProjectorSub, err := database.Collection(req.DB, &models.Meeting{}).SetFqids(projection.ContentObjectID).SetFields("reference_projector_id").SubscribeField(&referenceProjectorId)
+func CurrentSpeakerChyronSlideHandler(ctx context.Context, req *projectionRequest) (any, error) {
+	referenceProjectorId, err := req.Fetch.Meeting_ReferenceProjectorID(*req.ContentObjectID).Value(ctx)
 	if err != nil {
-		return nil, fmt.Errorf("failed to subscribe reference projector id: %w", err)
+		return nil, fmt.Errorf("could not load reference projector id %w", err)
 	}
 
-	var projector models.Projector
-	projectorQ := database.Collection(req.DB, &models.Projector{}).With("current_projection_ids", nil)
-	projectorSub, err := projectorQ.SubscribeOne(&projector)
+	losID, err := viewmodels.Projector_ListOfSpeakersID(ctx, req.Fetch, referenceProjectorId)
 	if err != nil {
-		return nil, fmt.Errorf("failed to subscribe reference projector: %w", err)
+		return nil, fmt.Errorf("could not load list of speakers id %w", err)
 	}
 
-	projectionsQ := projectorQ.GetSubquery("current_projection_ids")
-	projectionsQ.With("content_object_id", nil)
+	if losID == nil {
+		return nil, nil
+	}
 
-	var los models.ListOfSpeakers
-	losQ := database.Collection(req.DB, &models.ListOfSpeakers{}).With("speaker_ids", nil)
-	losQ.With("meeting_id", []string{"list_of_speakers_default_structure_level_time"})
-	losContentObjectQ := losQ.With("content_object_id", nil)
-	losContentObjectQ.With("agenda_item_id", nil)
-	speakersQ := losQ.GetSubquery("speaker_ids")
-	sllosQ := speakersQ.With("structure_level_list_of_speakers_id", nil)
-	sllosQ.With("structure_level_id", []string{"name"})
-	meetingUsersQ := speakersQ.With("meeting_user_id", nil)
-	meetingUsersQ.With("user_id", nil)
-	meetingUsersQ.With("structure_level_ids", []string{"name"})
-
-	losSub, err := losQ.SubscribeOne(&los)
+	losQ := req.Fetch.ListOfSpeakers(*losID)
+	los, err := losQ.Preload(losQ.SpeakerList().MeetingUser().User()).First(ctx)
 	if err != nil {
-		return nil, fmt.Errorf("failed to subscribe list of speakers: %w", err)
+		return nil, fmt.Errorf("could not load list of speakers: %w", err)
 	}
 
 	var options currentSpeakerChyronSlideOptions
-	if err := json.Unmarshal(projection.Options, &options); err != nil {
+	if err := json.Unmarshal(req.Projection.Options, &options); err != nil {
 		return nil, fmt.Errorf("could not parse slide options: %w", err)
 	}
 
-	go func() {
-		for {
-			select {
-			case <-ctx.Done():
-				refProjectorSub.Unsubscribe()
-				projectorSub.Unsubscribe()
-				losSub.Unsubscribe()
-				close(content)
-				return
-			case <-refProjectorSub.Channel:
-				if referenceProjectorId > 0 {
-					projectorQ.SetIds(referenceProjectorId)
-					if err := projectorSub.Reload(); err != nil {
-						log.Err(err).Msg("Reference projector load failed")
-					}
-				}
-			case <-projectorSub.Channel:
-				for _, p := range projector.CurrentProjections() {
-					if p.ContentObjectID == "" {
-						continue
-					}
-
-					losId := p.ContentObject().Get("list_of_speakers_id")
-					if losId != nil {
-						v := reflect.ValueOf(losId)
-						if v.Kind() == reflect.Ptr {
-							v = v.Elem()
-						}
-
-						losQ.SetIds(int(v.Int()))
-						if err := losSub.Reload(); err != nil {
-							log.Err(err).Msg("Reference projector load failed")
-						}
-						break
-					}
-				}
-			case <-losSub.Channel:
-				if los.ID != 0 {
-					content <- getSpeakerChyronSlideContent(&los, options)
-				} else {
-					content <- ""
-				}
-			}
-		}
-	}()
-
-	return content, nil
-}
-
-func getSpeakerChyronSlideContent(los *models.ListOfSpeakers, options currentSpeakerChyronSlideOptions) string {
-	tmpl, err := template.ParseFiles("templates/slides/current-speaker-chyron.html")
+	currentSpeaker, err := viewmodels.ListOfSpeakers_CurrentSpeaker(ctx, &los)
 	if err != nil {
-		log.Error().Err(err).Msg("could not load current-list-of-speakers template")
-		return ""
+		return nil, fmt.Errorf("could not get current speaker: %w", err)
 	}
 
-	var currentSpeaker *models.Speaker
-	for _, speaker := range los.Speakers() {
-		if speaker.IsCurrent() {
-			speechState := ""
-			if speaker.SpeechState != nil {
-				speechState = *speaker.SpeechState
+	slideSpeakerName := ""
+	slideStructureLevel := ""
+	slideAgendaItem := ""
+	if currentSpeaker != nil {
+		speakerName, err := viewmodels.Speaker_FullName(ctx, currentSpeaker)
+		if err != nil {
+			return nil, fmt.Errorf("could not get speaker name: %w", err)
+		}
+
+		if speakerName != nil {
+			slideSpeakerName = *speakerName
+
+			structureLevelDefaultTime, err := req.Fetch.Meeting_ListOfSpeakersDefaultStructureLevelTime(los.MeetingID).Value(ctx)
+			if err != nil {
+				return nil, fmt.Errorf("could not load ListOfSpeakersDefaultStructureLevelTime: %w", err)
 			}
 
-			if speechState == "interposed_question" {
-				currentSpeaker = speaker
-				break
+			if structureLevelDefaultTime > 0 {
+				structureLevel, err := viewmodels.Speaker_StructureLevelName(ctx, currentSpeaker)
+				if err != nil {
+					return nil, fmt.Errorf("could not get speaker structurelevels: %w", err)
+				}
+
+				if structureLevel != nil {
+					slideStructureLevel = *structureLevel
+				}
 			} else {
-				currentSpeaker = speaker
-			}
-		}
-	}
+				if meetingUser, isSet := currentSpeaker.MeetingUser.Value(); isSet {
+					structureLevels, err := viewmodels.MeetingUser_StructureLevelNames(ctx, &meetingUser)
+					if err != nil {
+						return nil, fmt.Errorf("could not load structure levels: %w", err)
+					}
 
-	speakerName := ""
-	structureLevel := ""
-	agendaItem := ""
-	if currentSpeaker != nil && currentSpeaker.MeetingUser() != nil {
-		u := currentSpeaker.MeetingUser().User()
-		speakerName = u.ShortName()
-
-		structureLevelDefaultTime := los.Meeting().ListOfSpeakersDefaultStructureLevelTime
-		if structureLevelDefaultTime != nil && *structureLevelDefaultTime > 0 {
-			sllos := currentSpeaker.StructureLevelListOfSpeakers()
-			if sllos != nil {
-				structureLevel = sllos.StructureLevel().Name
-			}
-		} else {
-			structureLevelNames := []string{}
-			for _, sl := range currentSpeaker.MeetingUser().StructureLevels() {
-				structureLevelNames = append(structureLevelNames, sl.Name)
+					slideStructureLevel = structureLevels
+				}
 			}
 
-			structureLevel = strings.Join(structureLevelNames, ", ")
-		}
-
-		if options.ChyronType == "new" && structureLevel != "" {
-			speakerName = fmt.Sprintf("%s, %s", speakerName, structureLevel)
+			if options.ChyronType == "new" && slideStructureLevel != "" {
+				slideSpeakerName = fmt.Sprintf("%s, %s", slideSpeakerName, slideStructureLevel)
+			}
 		}
 	}
 
 	// TODO: Also include agenda item number and number
-	coTitle := los.ContentObject().Get("title")
-	if coTitle != nil {
-		agendaItem = coTitle.(string)
-	}
+	// coTitle := los.ContentObject().Get("title")
+	// if coTitle != nil {
+	// agendaItem = coTitle.(string)
+	// }
 
-	slideData := map[string]interface{}{
+	return map[string]any{
 		"Options":        options,
-		"SpeakerName":    speakerName,
-		"StructureLevel": structureLevel,
-		"AgendaItem":     agendaItem,
-	}
-
-	var content bytes.Buffer
-	if err := tmpl.Execute(&content, slideData); err != nil {
-		log.Error().Err(err).Msg("could not execute current-list-of-speakers template")
-		return ""
-	}
-
-	return content.String()
+		"SpeakerName":    slideSpeakerName,
+		"StructureLevel": slideStructureLevel,
+		"AgendaItem":     slideAgendaItem,
+	}, nil
 }
