@@ -2,47 +2,33 @@ package database
 
 import (
 	"context"
-	"fmt"
 
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/redis/go-redis/v9"
+	"github.com/OpenSlides/openslides-go/datastore/dskey"
+	"github.com/OpenSlides/openslides-go/datastore/dsmodels"
+	"github.com/OpenSlides/openslides-go/datastore/dsrecorder"
 )
 
-type Datastore struct {
-	ctx    context.Context
-	pool   *pgxpool.Pool
-	redis  *redis.Client
-	change *changeListenerServer
+type dsChangeListener struct {
+	ctx     context.Context
+	keys    map[dskey.Key]struct{}
+	handler func()
 }
 
-func New(addr string, redisAddr string) (*Datastore, error) {
-	config, err := pgxpool.ParseConfig(addr)
-	if err != nil {
-		return nil, fmt.Errorf("parse config: %w", err)
+func (db *Datastore) NewContext(ctx context.Context, handler func(*dsmodels.Fetch)) {
+	recorder := dsrecorder.New(db.ds)
+	fetch := dsmodels.New(recorder)
+
+	handler(fetch)
+	listener := dsChangeListener{
+		ctx:  ctx,
+		keys: recorder.Keys(),
 	}
 
-	config.ConnConfig.DefaultQueryExecMode = pgx.QueryExecModeSimpleProtocol
-
-	pool, err := pgxpool.NewWithConfig(context.Background(), config)
-	if err != nil {
-		return nil, fmt.Errorf("creating connection pool: %w", err)
+	listener.handler = func() {
+		recorder.Reset()
+		handler(fetch)
+		listener.keys = recorder.Keys()
 	}
 
-	rdb := redis.NewClient(&redis.Options{
-		Addr: redisAddr,
-	})
-
-	ds := Datastore{pool: pool, redis: rdb, ctx: context.Background()}
-	go ds.setupRedisListener()
-
-	return &ds, nil
-}
-
-func Collection[T any, PT baseModelPtr[T]](ds *Datastore, coll *T) *query[T, PT] {
-	return &query[T, PT]{
-		collection: coll,
-		datastore:  ds,
-		subquerys:  map[string]*recursiveSubqueryList{},
-	}
+	db.dsListeners = append(db.dsListeners, &listener)
 }
