@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
+
+	"github.com/rs/zerolog/log"
 )
 
 func (s *projectorHttp) ProjectorSubscribeHandler() http.HandlerFunc {
@@ -12,20 +14,20 @@ func (s *projectorHttp) ProjectorSubscribeHandler() http.HandlerFunc {
 		id, err := strconv.Atoi(r.PathValue("id"))
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
-			fmt.Fprintln(w, `{"error": true, "msg": "Projector id invalid"}`)
+			writeResponse(w, `{"error": true, "msg": "Projector id invalid"}`)
 			return
 		}
 
 		content, err := s.projector.SubscribeProjectorContent(r.Context(), id)
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
-			fmt.Fprintln(w, `{"error": true, "msg": "Error reading projector content"}`)
+			writeResponse(w, `{"error": true, "msg": "Error reading projector content"}`)
 			return
 		}
 
 		if content == nil {
 			w.WriteHeader(http.StatusNotFound)
-			fmt.Fprintln(w, `{"error": true, "msg": "Projector not found"}`)
+			writeResponse(w, `{"error": true, "msg": "Projector not found"}`)
 			return
 		}
 
@@ -35,14 +37,14 @@ func (s *projectorHttp) ProjectorSubscribeHandler() http.HandlerFunc {
 			projectorContentRaw, err := s.projector.GetProjectorContent(id)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
-				fmt.Fprintln(w, `{"error": true, "msg": "Error reading projector content"}`)
+				writeResponse(w, `{"error": true, "msg": "Error reading projector content"}`)
 				return
 			}
 
 			currentContent, err := json.Marshal(projectorContentRaw)
 			if err != nil {
 				w.WriteHeader(http.StatusInternalServerError)
-				fmt.Fprintln(w, `{"error": true, "msg": "Error encoding projector content"}`)
+				writeResponse(w, `{"error": true, "msg": "Error encoding projector content"}`)
 				return
 			}
 			projectorContent = string(currentContent)
@@ -54,14 +56,18 @@ func (s *projectorHttp) ProjectorSubscribeHandler() http.HandlerFunc {
 		w.Header().Set("Connection", "keep-alive")
 
 		if needsInit {
-			fmt.Fprintf(w, "event: projector-replace\ndata: %s\n\n", projectorContent)
+			if _, err := fmt.Fprintf(w, "event: projector-replace\ndata: %s\n\n", projectorContent); err != nil {
+				log.Err(err).Msg("error sending event")
+			}
 		}
 		w.(http.Flusher).Flush()
 
 		for {
 			select {
 			case event := <-content:
-				fmt.Fprintf(w, "event: %s\ndata: %s\n\n", event.Event, event.Data)
+				if _, err := fmt.Fprintf(w, "event: %s\ndata: %s\n\n", event.Event, event.Data); err != nil {
+					log.Err(err).Msg("error sending event")
+				}
 				w.(http.Flusher).Flush()
 			case <-r.Context().Done():
 				return
