@@ -10,6 +10,7 @@ import (
 	"strings"
 
 	"github.com/OpenSlides/openslides-go/auth"
+	"github.com/OpenSlides/openslides-go/datastore/flow"
 	"github.com/OpenSlides/openslides-go/environment"
 	"github.com/OpenSlides/openslides-go/redis"
 	"github.com/OpenSlides/openslides-projector-service/pkg/database"
@@ -25,13 +26,14 @@ type projectorHttp struct {
 	ctx       context.Context
 	serverMux *http.ServeMux
 	db        *database.Datastore
+	ds        flow.Flow
 	projector *projector.ProjectorPool
 	cfg       ProjectorConfig
 	auth      *auth.Auth
 }
 
-func New(ctx context.Context, cfg ProjectorConfig, serverMux *http.ServeMux, db *database.Datastore) {
-	projectorPool := projector.NewProjectorPool(ctx, db)
+func New(ctx context.Context, cfg ProjectorConfig, serverMux *http.ServeMux, db *database.Datastore, ds flow.Flow) {
+	projectorPool := projector.NewProjectorPool(ctx, db, ds)
 
 	lookup := new(environment.ForProduction)
 	redis := redis.New(lookup)
@@ -48,11 +50,18 @@ func New(ctx context.Context, cfg ProjectorConfig, serverMux *http.ServeMux, db 
 		ctx:       ctx,
 		serverMux: serverMux,
 		db:        db,
+		ds:        ds,
 		projector: projectorPool,
 		auth:      authService,
 		cfg:       cfg,
 	}
 	handler.registerRoutes(cfg)
+}
+
+func writeResponse(w http.ResponseWriter, resp string) {
+	if _, err := fmt.Fprintln(w, resp); err != nil {
+		log.Err(err).Msg("writing response")
+	}
 }
 
 func (s *projectorHttp) registerRoutes(cfg ProjectorConfig) {
@@ -66,14 +75,14 @@ func authMiddleware(next http.Handler, auth *auth.Auth, cfg ProjectorConfig) htt
 		ctx, err := auth.Authenticate(w, r)
 		if err != nil {
 			w.WriteHeader(http.StatusUnauthorized)
-			fmt.Fprintln(w, `{"error": true, "msg": "authenticate request failed"}`)
+			writeResponse(w, `{"error": true, "msg": "authenticate request failed"}`)
 			return
 		}
 
 		id, err := strconv.Atoi(r.PathValue("id"))
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
-			fmt.Fprintln(w, `{"error": true, "msg": "Projector id invalid"}`)
+			writeResponse(w, `{"error": true, "msg": "Projector id invalid"}`)
 			return
 		}
 
@@ -83,7 +92,7 @@ func authMiddleware(next http.Handler, auth *auth.Auth, cfg ProjectorConfig) htt
 		restrictUrl := fmt.Sprintf("%s?user_id=%d&single=1", cfg.RestricterUrl, userID)
 		req, err := http.NewRequest("POST", restrictUrl, bytes.NewReader(body))
 		if err != nil {
-			fmt.Fprintln(w, `{"error": true, "msg": "creating restriction request failed"}`)
+			writeResponse(w, `{"error": true, "msg": "creating restriction request failed"}`)
 			return
 		}
 
@@ -95,21 +104,25 @@ func authMiddleware(next http.Handler, auth *auth.Auth, cfg ProjectorConfig) htt
 		resp, err := client.Do(req)
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
-			fmt.Fprintln(w, `{"error": true, "msg": "restriction request failed"}`)
+			writeResponse(w, `{"error": true, "msg": "restriction request failed"}`)
 			return
 		}
 
 		if resp.StatusCode != http.StatusOK {
 			w.WriteHeader(resp.StatusCode)
-			fmt.Fprintln(w, `{"error": true, "msg": "restriction request failed"}`)
+			writeResponse(w, `{"error": true, "msg": "restriction request failed"}`)
 			return
 		}
-		defer resp.Body.Close()
+
 		b, err := io.ReadAll(resp.Body)
 		if err != nil || !strings.Contains(string(b), fmt.Sprintf(`"projector/%d/id":%d`, id, id)) {
 			w.WriteHeader(http.StatusUnauthorized)
-			fmt.Fprintln(w, `{"error": true, "msg": "permissions denied"}`)
+			writeResponse(w, `{"error": true, "msg": "permissions denied"}`)
 			return
+		}
+
+		if err := resp.Body.Close(); err != nil {
+			log.Err(err).Msg("error closing response body")
 		}
 
 		next.ServeHTTP(w, r.WithContext(ctx))
