@@ -1,7 +1,7 @@
 import * as esbuild from 'esbuild';
 
 await esbuild.build({
-  entryPoints: ['src/projector.js', 'src/projector.css', 'src/slide/*.css', 'src/slide/*.js'],
+  entryPoints: ['src/projector.js', 'src/projector.css', 'src/projector-page.css', 'src/slide/*.css', 'src/slide/*.js'],
   bundle: true,
   minify: true,
   sourcemap: true,
